+++ src/components/modals/SwapVerifyTermsModal.tsx
@@ -8,8 +8,9 @@ import { lstrings } from '../../locales/strings'
 import { getSwapPluginIconUri } from '../../util/CdnUris'
 import { Airship } from '../services/AirshipInstance'
 import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
+import { Paragraph } from '../themed/EdgeText'
 import { MainButton } from '../themed/MainButton'
-import { ModalMessage, ModalTitle } from '../themed/ModalParts'
+import { ModalTitle } from '../themed/ModalParts'
 import { ModalUi4 } from '../ui4/ModalUi4'
 
 interface TermsUri {
@@ -69,7 +70,7 @@ function SwapVerifyTermsModal(props: Props) {
       }
       onCancel={() => bridge.resolve(false)}
     >
-      <ModalMessage>{lstrings.swap_terms_statement}</ModalMessage>
+      <Paragraph>{lstrings.swap_terms_statement}</Paragraph>
       <MainButton label={lstrings.swap_terms_accept_button} marginRem={1} onPress={() => bridge.resolve(true)} />
       <MainButton label={lstrings.swap_terms_reject_button} marginRem={1} type="secondary" onPress={() => bridge.resolve(false)} />
       <View style={styles.linkContainer}>